requirements-dev.txt
blue
faker
ipython
isort
pip-tools
pre-commit
pytest
pytest-cov
pytest-django
pytest-sugar
ruff
taskipy
